requirements.txt
# Core dependencies
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10
pydantic==2.5.0
python-dotenv==1.0.0

# Scientific computing
numpy==1.24.3
scipy==1.11.4
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0

# MD and materials science
ase==3.22.1
pymatgen==2023.8.10
crystal-toolkit==2023.8.10
pymatgen-analysis-diffusion==2022.0.0
pymatgen-analysis-structure-matcher==2023.8.10

# LAMMPS interface
lammps-cython==1.0.0
lammps==2023.08.02

# ML and AI
torch==2.1.0
scikit-learn==1.3.2
tensorflow==2.15.0
transformers==4.35.2

# Database APIs
requests==2.31.0
aiohttp==3.9.0
mp-api==0.33.0

# Utilities
tqdm==4.66.1
pyyaml==6.0.1
h5py==3.9.0
netcdf4==1.6.4
pymatgen-io-lammps==0.0.1

# Development
pytest==7.4.3
black==23.11.0
flake8==6.1.0
jupyter==1.0.0

# GUI Requirements for Materials AI Agent
streamlit>=1.28.0
plotly>=5.17.0
pandas>=2.0.3
numpy>=1.24.3
matplotlib>=3.7.2
scipy>=1.11.4

# Core dependencies (if not already installed)
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
pydantic>=2.5.0
python-dotenv>=1.0.0

# Scientific computing
ase>=3.22.1
pymatgen>=2023.8.10
mp-api>=0.33.0

# Optional for enhanced functionality
scikit-learn>=1.3.2
seaborn>=0.12.2

